test: package manager check
