+++ pnpm/test/install/misc.ts
@@ -248,81 +248,6 @@ test('`pnpm -r add` should fail if no package name was provided', () => {
   expect(stdout.toString()).toContain('`pnpm add` requires the package name')
 })
 
-test('install should fail if the used pnpm version does not satisfy the pnpm version specified in engines', async () => {
-  prepare({
-    name: 'project',
-    version: '1.0.0',
-
-    engines: {
-      pnpm: '99999',
-    },
-  })
-
-  const { status, stdout } = execPnpmSync(['install'])
-
-  expect(status).toBe(1)
-  expect(stdout.toString()).toContain('Your pnpm version is incompatible with')
-})
-
-test('install should not fail if the used pnpm version does not satisfy the pnpm version specified in packageManager', async () => {
-  prepare({
-    name: 'project',
-    version: '1.0.0',
-
-    packageManager: 'pnpm@0.0.0',
-  })
-
-  expect(execPnpmSync(['install', '--config.manage-package-manager-versions=false']).status).toBe(0)
-
-  const { status, stderr } = execPnpmSync(['install', '--config.manage-package-manager-versions=false', '--config.package-manager-strict-version=true'])
-
-  expect(status).toBe(1)
-  expect(stderr.toString()).toContain('This project is configured to use v0.0.0 of pnpm. Your current pnpm is')
-})
-
-test('install should fail if the project requires a different package manager', async () => {
-  prepare({
-    name: 'project',
-    version: '1.0.0',
-
-    packageManager: 'yarn@4.0.0',
-  })
-
-  const { status, stderr } = execPnpmSync(['install', '--config.manage-package-manager-versions=true'])
-
-  expect(status).toBe(1)
-  expect(stderr.toString()).toContain('This project is configured to use yarn')
-
-  expect(execPnpmSync(['install', '--config.package-manager-strict=false']).status).toBe(0)
-})
-
-test('install should not fail for packageManager field with hash', async () => {
-  const versionProcess = execPnpmSync(['--version'])
-  const pnpmVersion = versionProcess.stdout.toString().trim()
-
-  prepare({
-    name: 'project',
-    version: '1.0.0',
-
-    packageManager: `pnpm@${pnpmVersion}+sha256.123456789`,
-  })
-
-  const { status } = execPnpmSync(['install'])
-  expect(status).toBe(0)
-})
-
-test('install should not fail for packageManager field with url', async () => {
-  prepare({
-    name: 'project',
-    version: '1.0.0',
-
-    packageManager: 'pnpm@https://github.com/pnpm/pnpm',
-  })
-
-  const { status } = execPnpmSync(['install'])
-  expect(status).toBe(0)
-})
-
 test('engine-strict=false: install should not fail if the used Node version does not satisfy the Node version specified in engines', async () => {
   prepare({
     name: 'project',
